product-bg-removal/requirements.txt
fastapi
uvicorn[standard]
rembg
Pillow
python-multipart
